Fix up test_validate to use conftest configs
Aiming for single-source-of-truth to be central conftest,
and for those bits of config to be consistent project elements.

+++ tests/data_layer/test_data_handle.py
@@ -66,7 +66,7 @@ def mock_store(sample_dimensions, get_sector_model, empty_store):
         'outputs': [
             {
                 'name': 'test',
-                'dtype': 'float',
+                'dtype': 'int',
                 'dims': ['lad', 'annual'],
                 'absolute_range': [0, int(1e12)],
                 'expected_range': [0, 100000],
@@ -74,7 +74,7 @@ def mock_store(sample_dimensions, get_sector_model, empty_store):
             },
             {
                 'name': 'test_alt_dims',
-                'dtype': 'float',
+                'dtype': 'int',
                 'dims': ['lad', 'annual'],
                 'absolute_range': [0, int(1e12)],
                 'expected_range': [0, 100000],
@@ -95,7 +95,7 @@ def mock_store(sample_dimensions, get_sector_model, empty_store):
         'inputs': [
             {
                 'name': 'test',
-                'dtype': 'float',
+                'dtype': 'int',
                 'dims': ['lad', 'annual'],
                 'absolute_range': [0, int(1e12)],
                 'expected_range': [0, 100000],
@@ -105,7 +105,7 @@ def mock_store(sample_dimensions, get_sector_model, empty_store):
         'outputs': [
             {
                 'name': 'test',
-                'dtype': 'float',
+                'dtype': 'int',
                 'dims': ['lad', 'annual'],
                 'absolute_range': [0, int(1e12)],
                 'expected_range': [0, 100000],
